change ids to 64 bit

+++ src/request.rs
@@ -1,6 +1,6 @@
 #[derive(Clone, Debug, Serialize)]
 pub struct Message {
-    pub chat_id: i32,
+    pub chat_id: i64,
     pub text: String,
     //TODO
 }
