+++ src/response.rs
@@ -1,6 +1,6 @@
 #[derive(Clone, Debug, Deserialize)]
 pub struct Message {
-    pub message_id: i32,
+    pub message_id: i64,
     pub from: User,
     pub date: i32,
     pub chat: Chat,
@@ -8,7 +8,7 @@ pub struct Message {
 }
 #[derive(Clone, Debug, Deserialize)]
 pub struct User {
-    pub id: i32,
+    pub id: i64,
     pub first_name: String,
     pub last_name: Option<String>,
     pub username: Option<String>,
@@ -16,7 +16,7 @@ pub struct User {
 }
 #[derive(Clone, Debug, Deserialize)]
 pub struct Chat {
-    pub id: i32,
+    pub id: i64,
     #[serde(rename="type")]
     pub chat_type: ChatType,
     pub title: Option<String>,
